% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{read_spec_file}
\alias{read_spec_file}
\title{Title}
\usage{
read_spec_file(file, folder = NULL)
}
\arguments{
\item{folder}{}
}
\description{
Title
}
\examples{
folder<-system.file("examples", package = "rjd3tramoseatsx")
file<-"example.7"
q<-read_spec_file(file, folder)
rjd3tramoseats::tramoseats_fast(q$series, q$spec, q$context)
}
